Move runJob to utils.js, store the id in a constant
Create constants RUN_JOB_COMMAND
so 2 files can use it.
