+++ src/Job.ts
@@ -1,5 +1,6 @@
 import * as vscode from 'vscode';
 import * as path from 'path';
+import { RUN_JOB_COMMAND } from './constants';
 
 export class Job extends vscode.TreeItem {
   constructor(
@@ -12,7 +13,7 @@ export class Job extends vscode.TreeItem {
     this.tooltip = tooltip;
     this.command = {
       title: label,
-      command: 'localci.runAction',
+      command: RUN_JOB_COMMAND,
       tooltip,
       arguments: [label],
     };
